@@ -12,32 +12,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-mdl/core": "^4.0.5",
-    "@angular/animations": "^4.1.3",
-    "@angular/common": "^4.1.3",
-    "@angular/compiler": "^4.1.3",
-    "@angular/core": "^4.1.3",
-    "@angular/forms": "^4.1.3",
-    "@angular/http": "^4.1.3",
-    "@angular/platform-browser": "^4.1.3",
-    "@angular/platform-browser-dynamic": "^4.1.3",
-    "@angular/router": "^4.1.3",
-    "@swimlane/ngx-datatable": "^9.2.0",
-    "core-js": "^2.4.1",
-    "nedb": "^1.8.0",
-    "ngx-electron": "1.0.2",
-    "rxjs": "^5.4.0",
-    "typescript": "^2.3.3",
-    "zone.js": "^0.8.11"
+    "@angular/animations": "^5.2.9",
+    "@angular/cdk": "^5.2.4",
+    "@angular/common": "^5.2.9",
+    "@angular/compiler": "^5.2.9",
+    "@angular/core": "^5.2.9",
+    "@angular/forms": "^5.2.9",
+    "@angular/material": "^5.2.4",
+    "@angular/platform-browser": "^5.2.9",
+    "@angular/platform-browser-dynamic": "^5.2.9",
+    "chart.js": "^2.7.2",
+    "ngx-electron": "^1.0.4",
+    "rxjs": "^5.5.7",
+    "zone.js": "^0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "^1.1.0",
-    "@angular/compiler-cli": "^4.1.3",
-    "@types/nedb": "^1.8.3",
-    "@types/node": "~7.0.22",
-    "codelyzer": "~3.0.1",
-    "concurrently": "^3.4.0",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.4.2"
+    "@angular/cli": "^1.7.3",
+    "@angular/compiler-cli": "^5.2.9",
+    "@angular/language-service": "^5.2.9",
+    "typescript": "^2.6.2"
   }
 }
